chore: extract types and common code
