+++ js/ViewPagerAndroid.android.js
@@ -8,7 +8,15 @@
  * @flow strict-local
  */
 
+
 'use strict';
+
+import type {
+  PageScrollEvent,
+  PageScrollStateChangedEvent,
+  PageSelectedEvent,
+  ViewPagerProps
+} from "./types";
 
 const React = require('react');
 const ReactNative = require('react-native');
@@ -16,8 +24,7 @@ const {UIManager} = ReactNative;
 
 const dismissKeyboard = require('react-native/Libraries/Utilities/dismissKeyboard');
 
-import type {SyntheticEvent} from 'react-native/Libraries/Types/CoreEventTypes';
-import type {ViewStyleProp} from 'react-native/Libraries/StyleSheet/StyleSheet';
+import {childrenWithOverriddenStyle} from "./utils";
 
 const NativeAndroidViewPager = require('./AndroidViewPagerNativeComponent');
 
@@ -30,100 +37,6 @@ function getViewManagerConfig(viewManagerName) {
   }
   return UIManager.getViewManagerConfig(viewManagerName);
 }
-
-type PageScrollState = 'idle' | 'dragging' | 'settling';
-
-type PageScrollEvent = SyntheticEvent<
-  $ReadOnly<{|
-    position: number,
-    offset: number,
-  |}>,
->;
-
-type PageScrollStateChangedEvent = SyntheticEvent<
-  $ReadOnly<{|
-    pageScrollState: PageScrollState,
-  |}>,
->;
-
-type PageSelectedEvent = SyntheticEvent<
-  $ReadOnly<{|
-    position: number,
-  |}>,
->;
-
-type ViewPagerScrollState = $Enum<{
-  idle: string,
-  dragging: string,
-  settling: string,
-}>;
-
-type Props = $ReadOnly<{|
-  /**
-   * Index of initial page that should be selected. Use `setPage` method to
-   * update the page, and `onPageSelected` to monitor page changes
-   */
-  initialPage?: ?number,
-
-  /**
-   * Executed when transitioning between pages (ether because of animation for
-   * the requested page change or when user is swiping/dragging between pages)
-   * The `event.nativeEvent` object for this callback will carry following data:
-   *  - position - index of first page from the left that is currently visible
-   *  - offset - value from range [0,1) describing stage between page transitions.
-   *    Value x means that (1 - x) fraction of the page at "position" index is
-   *    visible, and x fraction of the next page is visible.
-   */
-  onPageScroll?: ?(e: PageScrollEvent) => void,
-
-  /**
-   * Function called when the page scrolling state has changed.
-   * The page scrolling state can be in 3 states:
-   * - idle, meaning there is no interaction with the page scroller happening at the time
-   * - dragging, meaning there is currently an interaction with the page scroller
-   * - settling, meaning that there was an interaction with the page scroller, and the
-   *   page scroller is now finishing it's closing or opening animation
-   */
-  onPageScrollStateChanged?: ?(e: PageScrollStateChangedEvent) => void,
-
-  /**
-   * This callback will be called once ViewPager finish navigating to selected page
-   * (when user swipes between pages). The `event.nativeEvent` object passed to this
-   * callback will have following fields:
-   *  - position - index of page that has been selected
-   */
-  onPageSelected?: ?(e: PageSelectedEvent) => void,
-
-  /**
-   * Blank space to show between pages. This is only visible while scrolling, pages are still
-   * edge-to-edge.
-   */
-  pageMargin?: ?number,
-
-  /**
-   * Whether enable showing peekFraction or not. If this is true, the preview of
-   * last and next page will show in current screen. Defaults to false.
-   */
-
-  peekEnabled?: ?boolean,
-
-  /**
-   * Determines whether the keyboard gets dismissed in response to a drag.
-   *   - 'none' (the default), drags do not dismiss the keyboard.
-   *   - 'on-drag', the keyboard is dismissed when a drag begins.
-   */
-  keyboardDismissMode?: ?('none' | 'on-drag'),
-
-  /**
-   * When false, the content does not scroll.
-   * The default value is true.
-   */
-  scrollEnabled?: ?boolean,
-
-  children?: React.Node,
-
-  style?: ?ViewStyleProp,
-|}>;
 
 /**
  * Container that allows to flip left and right between child views. Each
@@ -167,7 +80,7 @@ type Props = $ReadOnly<{|
  * ```
  */
 
-class ViewPagerAndroid extends React.Component<Props> {
+class ViewPagerAndroid extends React.Component<ViewPagerProps> {
   componentDidMount() {
     if (this.props.initialPage != null) {
       this.setPageWithoutAnimation(this.props.initialPage);
@@ -178,47 +91,6 @@ class ViewPagerAndroid extends React.Component<Props> {
    * when making Flow check .android.js files. */
   getInnerViewNode = (): ReactComponent => {
     return this.refs[VIEWPAGER_REF].getInnerViewNode();
-  };
-
-  /* $FlowFixMe(>=0.78.0 site=react_native_android_fb) This issue was found
-   * when making Flow check .android.js files. */
-  _childrenWithOverridenStyle = (): Array => {
-    // Override styles so that each page will fill the parent. Native component
-    // will handle positioning of elements, so it's not important to offset
-    // them correctly.
-    return React.Children.map(this.props.children, function(child) {
-      if (!child) {
-        return null;
-      }
-      const newProps = {
-        ...child.props,
-        style: [
-          child.props.style,
-          {
-            position: 'absolute',
-            left: 0,
-            top: 0,
-            right: 0,
-            bottom: 0,
-            width: undefined,
-            height: undefined,
-          },
-        ],
-        collapsable: false,
-      };
-      if (
-        child.type &&
-        child.type.displayName &&
-        child.type.displayName !== 'RCTView' &&
-        child.type.displayName !== 'View'
-      ) {
-        console.warn(
-          'Each ViewPager child must be a <View>. Was ' +
-            child.type.displayName,
-        );
-      }
-      return React.createElement(child.type, newProps);
-    });
   };
 
   _onPageScroll = (e: PageScrollEvent) => {
@@ -276,7 +148,7 @@ class ViewPagerAndroid extends React.Component<Props> {
         onPageScroll={this._onPageScroll}
         onPageScrollStateChanged={this._onPageScrollStateChanged}
         onPageSelected={this._onPageSelected}
-        children={this._childrenWithOverridenStyle()}
+        children={childrenWithOverriddenStyle(this.props.children)}
       />
     );
   }
