+++ js/ViewPagerAndroid.ios.js
@@ -5,42 +5,21 @@
  * LICENSE file in the root directory of this source tree.
  *
  * @format
- * @flow
+ * @flow strict-local
  */
+
 
 'use strict';
 
+import type {ViewPagerProps} from "./types";
+
 const VIEWPAGER_REF = 'viewPager';
 
+import * as React from "react";
 import ReactNative, {UIManager} from 'react-native'
-import type {ViewStyleProp} from 'react-native/Libraries/StyleSheet/StyleSheet';
-import React from "react";
 import NativeAndroidViewPager from './AndroidViewPagerNativeComponent';
+import {childrenWithOverriddenStyle} from "./utils";
 
-type TransitionStyle = 'scroll' | 'curl'
-type Orientation = 'horizontal' | 'vertical'
-
-type PageSelectedEvent = SyntheticEvent<
-  $ReadOnly<{|
-    position: number,
-  |}>,
->;
-
-type Props = $ReadOnly<{|
-    /**
-     * Index of initial page that should be selected. Use `setPage` method to
-     * update the page, and `onPageSelected` to monitor page changes
-     */
-    onPageSelected?: ?(e: PageSelectedEvent) => void,
-    orientation?: Orientation,
-    transitionStyle?: TransitionStyle,
-    pageMargin?: number,
-    scrollEnabled?: boolean,
-    initialPage?: number,
-    style?: ?ViewStyleProp,
-    keyboardDismissMode?: ?('none' | 'on-drag'),
-  |}>;
-  
   /**
    * Container that allows to flip left and right between child views. Each
    * child view of the `ViewPagerAndroid` will be treated as a separate page
@@ -91,47 +70,7 @@ function getViewManagerConfig(viewManagerName) {
   return UIManager.getViewManagerConfig(viewManagerName);
 }
 
-class ViewPagerAndroid extends React.Component<Props> {
-  
-  _childrenWithOverridenStyle = (): Array => {
-    // Override styles so that each page will fill the parent. Native component
-    // will handle positioning of elements, so it's not important to offset
-    // them correctly.
-    return React.Children.map(this.props.children, function(child) {
-      if (!child) {
-        return null;
-      }
-      const newProps = {
-        ...child.props,
-        style: [
-          child.props.style,
-          {
-            position: 'absolute',
-            left: 0,
-            top: 0,
-            right: 0,
-            bottom: 0,
-            width: undefined,
-            height: undefined,
-          },
-        ],
-        collapsable: false,
-      };
-      if (
-        child.type &&
-        child.type.displayName &&
-        child.type.displayName !== 'RCTView' &&
-        child.type.displayName !== 'View'
-      ) {
-        console.warn(
-          'Each ViewPager child must be a <View>. Was ' +
-            child.type.displayName,
-        );
-      }
-      return React.createElement(child.type, newProps);
-    });
-  };
-  
+class ViewPagerAndroid extends React.Component<ViewPagerProps> {
   setPage = (selectedPage: number) => {
       UIManager.dispatchViewManagerCommand(
         ReactNative.findNodeHandle(this),
@@ -139,7 +78,7 @@ class ViewPagerAndroid extends React.Component<Props> {
         [selectedPage,true],
       );
     };
-  
+
     setPageWithoutAnimation = (selectedPage: number) => {
       UIManager.dispatchViewManagerCommand(
         ReactNative.findNodeHandle(this),
@@ -153,7 +92,7 @@ class ViewPagerAndroid extends React.Component<Props> {
         <NativeAndroidViewPager
           {...this.props}
           ref={VIEWPAGER_REF}
-          children={this._childrenWithOverridenStyle()}
+          children={childrenWithOverriddenStyle(this.props.children)}
         />
       );
     }
